# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	bridge/evm/broadcast/deploy_bridge.s.sol/12/run-latest.json
#	crates/sui-bridge/src/server/mod.rs
#	crates/sui-e2e-tests/tests/reconfiguration_tests.rs
#	crates/sui-framework-snapshot/bytecode_snapshot/79/0x0000000000000000000000000000000000000000000000000000000000000001
#	crates/sui-framework-snapshot/bytecode_snapshot/79/0x0000000000000000000000000000000000000000000000000000000000000002
#	crates/sui-framework-snapshot/bytecode_snapshot/79/0x0000000000000000000000000000000000000000000000000000000000000003
#	crates/sui-framework-snapshot/bytecode_snapshot/79/0x000000000000000000000000000000000000000000000000000000000000000b
#	crates/sui-framework-snapshot/bytecode_snapshot/79/0x00000000000000000000000000000000000000000000000000000000000000c8
#	crates/sui-framework/docs/bridge/bridge.md
#	crates/sui-framework/packages/bridge/tests/bridge_env.move
#	crates/sui-framework/packages/bridge/tests/bridge_tests.move
#	crates/sui-protocol-config/src/lib.rs
#	crates/sui-rosetta/tests/custom_coins/test_coin/Move.toml
#	crates/sui-rosetta/tests/custom_coins/test_coin_no_symbol/Move.toml
#	crates/sui-swarm-config/tests/snapshots/snapshot_tests__genesis_config_snapshot_matches.snap
#	crates/sui-swarm-config/tests/snapshots/snapshot_tests__populated_genesis_snapshot_matches-2.snap
#	rust-toolchain.toml
#	scripts/simtest/cargo-simtest

+++ crates/sui-framework/packages/bridge/sources/bridge_fee.move
@@ -425,7 +425,7 @@ module bridge::bridge_fee{
 
     fun calculate_fee(amount: u64,fee_rate: u64) : u64 {
         assert!(fee_rate < FEE_RATE_PRECISION, EBridgeFeeSettingWrong);
-        (amount * fee_rate) / FEE_RATE_PRECISION
+        (((amount as u128)*( fee_rate as u128)) /(FEE_RATE_PRECISION as u128)) as u64
     }
 
 }
